Modify: 예약페이지 상태값 큰 컨테이너에서 관리
